refactor(bot): move SeriesParser modules to its own package
