+++ fazcord/series_parser/_base_series_parser.py
